Update README.md
Short section added about how the database is organized. This readme could be further expanded to include more of the overview information currently covered in the Lemmy post, and updated to reflect the most up to date methods for submission now that we have this github helping to gather thoughts and feedback.